Add GitHub Actions for zip release and reformat the directory structure. (#13)
* Add GitHub Actions for zip release and reformat the directory structure.

Introduce a workflow to automate zip releases upon pushing to the main branch, and add relevant artifacts to the release. Also, update organization IDs in teams.json for consistency and added necessary entries to .gitignore.

* Add proper CDS format and add a zip compatible with domjudge import script to releases.

* Update workflow to add organizations.json to cds.zip, rename cds.zip to cdf.zip, and enhance organizations.json

* Rename zip file from cpf.zip to contest-package.zip

This renaming is applied in the GitHub Actions workflow for clarity and consistency. The updated filenames are reflected in both the zip creation and release files steps.

* rename the 512px logo to the correct format